108. Convert Sorted Array to Binary Search Tree/first/SortedArrayToBST.java
package first;

import common.TreeNode;

public class SortedArrayToBST {
    public TreeNode sortedArrayToBST(int[] nums) {
        return recur(nums, 0, nums.length - 1);
    }

    private TreeNode recur(int[] nums, int lo, int hi) {
        if (lo > hi) return null;
        if (lo == hi) return new TreeNode(nums[lo]);

        int mid = lo + (hi - lo) / 2;
        TreeNode root = new TreeNode(nums[mid]);
        root.left = recur(nums, lo, mid - 1);
        root.right = recur(nums, mid + 1, hi);
        return root;
    }
}
